Move CallStateExtending to its own file
